# compiletest: Default to one CGU when compiling tests.