@aidanhs, maybe you want to give this a try. It reduced the execution time for `run-pass` from 8 min 30s to 7 min 5s on my machine (though that was just one test run, so ymmv).

r? @aidanhs