@kennytm 2018-01-26
(I've restarted the Travis CI because I find that the timings in this PR is slower than master or another randomly picked PR in every step, suggesting Travis had assigned a slower machine here. Restarting to try to get a fairer compare result.)