@aidanhs 2018-03-05
I tried this locally via `./stamp ./x.py test --stage 1 src/test/run-pass` on 16.04, 4C/8T CPU
 - with `--disable-optimize` (the default), the tests went from 221.195s to 192.175s
 - with `--enable-optimize`, the tests went from 271.008s to 198.000s

Definitely seems to make a difference. I had a thought though - codegen-units allows rustc itself to parallelise, and rust tests are run in parallel. Doesn't this mean that we're parallelising squared? Watching the load avg on my machine seems to confirm this - it hovers around 8 with this PR enabled (i.e. with codegen-units=1), but becomes at least double without it. Will do some more digging.